refactor core handlers
